0.2.2 Release * Get the Win32 executable functional again by heavily patching Urho (OpenGL only) * Docker container build scripts and methods for compiling the targets added to the tree * Simple preferences dialog added with some minimal options for those who didn't know   about the command line options for reseting the wizardry path, mainly. * Switch on Fractional scaling in HighDPI, which might improve the visuals * Add an id column to the fact editor, since the numbers get used a lot. * BUGFIX: Original games under Wizardry 28 have their own headaches, because certain   elements like spells are still in the original format, but everything else changes -   no clean way to detect this situation but can support it based on name if ParallelWorlds   active. * BUGFIX: Try to resolve some of the problems surrounding dead characters in a party being   edited.

+++ src/DialogAboutUrho3D.cpp
@@ -909,7 +909,7 @@ void DialogAboutUrho3D::ColouriseLetters( int x, int y)
             float distance = sqrtf((pos.x_ - mouseLoc.x_) * (pos.x_ - mouseLoc.x_) +
                                    (pos.y_ - mouseLoc.y_) * (pos.y_ - mouseLoc.y_));
 
-            for (oldcol=0; oldcol < sizeof(flame) / sizeof(Color); oldcol++)
+            for (oldcol=0; oldcol < (int)(sizeof(flame) / sizeof(Color)); oldcol++)
             {
                 if ((fabs(col.r_ - flame[oldcol].r_) < EPSILON) &&
                     (fabs(col.g_ - flame[oldcol].g_) < EPSILON) &&
